# Optimize the optimize command